requirements.txt
streamlit
regex
sentence_transformers
scipy
tqdm
plotly
wordcloud
matplotlib
spacy>=2.2.0,<3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz
bert-extractive-summarizer